public/requirements.txt
flask
yt-dlp
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
youtube-transcript-api
python-dotenv
gunicorn
gevent
wheel
